entrega dh

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ class Postcontroller extends Controller
 {
   public function getHome()
   {
-    $posts = Post::all();
+    $posts = Post::orderBy('created_at', 'desc')-> get();
     return view('home', ['posts'=> $posts]);
   }
 
@@ -39,7 +39,30 @@ class Postcontroller extends Controller
   public function getDeletePost($post_id)
   {
     $post = Post::where('id', $post_id)->first();
+    if(Auth::user() != $post->user){
+      return redirect()->back();
+    }
     $post->delete();
     return redirect()->route('home')->with(['message' => 'Te fuiste DE-LE-TEA-DO']);
   }
+
+
+  public function postEditPost(Request $request)
+  {
+    $this->validate($request, [
+      'body' => 'required'
+    ]);
+   $post = Post::find($request['postId']);
+   if(Auth::user() != $post->user){
+     return redirect()->back();
+   }
+   $post->body = $request['body'];
+   $post->update();
+   return response()->json(['new_body' => $post->body],200);
+ }
+
+
+
+
+
 }
